refactor(surveybar): update based on comments in PR
